+++ packages/cadl-ranch-specs/http/azure/resource-manager/common-type/managed-identity/main.tsp
@@ -16,7 +16,7 @@ using TypeSpec.OpenAPI;
 @service
 @versioned(Versions)
 @doc("Arm Managed Identity Provider management API.")
-namespace Azure.ResourceManager.CommonTypes;
+namespace Azure.ResourceManager.CommonType;
 
 @doc("Azure API versions.")
 enum Versions {
@@ -46,17 +46,17 @@ model ManagedIdentityTrackedResourceProperties {
 }
 
 @armResourceOperations
-interface ManagedIdentities {
+interface ManagedIdentity {
   @scenario
   @scenarioDoc("""
     Resource GET operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/identity",
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonType.ManagedIdentity/managedIdentityTrackedResources/identity",
     Expected query parameter: api-version=2023-12-01-preview
     
     Expected response body:
     ```json
     {
-      "id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/identity",
+      "id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonType.ManagedIdentity/managedIdentityTrackedResources/identity",
       "location": "eastus",
       "tags": {
         "tagKey1": "tagValue1"
@@ -77,7 +77,7 @@ interface ManagedIdentities {
   @scenario
   @scenarioDoc("""
     Resource PUT operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/identity",
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonType.ManagedIdentity/managedIdentityTrackedResources/identity",
     Expected query parameter: api-version=2023-12-01-preview
     Expected request body:
     ```json
@@ -95,7 +95,7 @@ interface ManagedIdentities {
      Expected response body:
     ```json
     {
-      "id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/identity",
+      "id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonType.ManagedIdentity/managedIdentityTrackedResources/identity",
       "location": "eastus",
       "tags": {
         "tagKey1": "tagValue1"
@@ -116,7 +116,7 @@ interface ManagedIdentities {
   @scenario
   @scenarioDoc("""
     Resource PATCH operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/identity",
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonType.ManagedIdentity/managedIdentityTrackedResources/identity",
     Expected query parameter: api-version=2023-12-01-preview
     Expected request body:
     ```json
@@ -132,7 +132,7 @@ interface ManagedIdentities {
      Expected response body:
     ```json
     {
-      "id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/identity",
+      "id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.CommonType.ManagedIdentity/managedIdentityTrackedResources/identity",
       "location": "eastus",
       "tags": {
         "tagKey1": "tagValue1"
