+++ packages/cadl-ranch-specs/http/azure/resource-manager/resource/toplevel.tsp
@@ -9,7 +9,7 @@ using Azure.Core;
 using Azure.ResourceManager;
 using TypeSpec.OpenAPI;
 
-namespace Azure.ResourceManager.Resources;
+namespace Azure.ResourceManager.Resource;
 
 @resource("topLevelTrackedResources")
 model TopLevelTrackedResource is TrackedResource<TopLevelTrackedResourceProperties> {
@@ -41,17 +41,17 @@ model NotificationDetails {
 }
 
 @armResourceOperations
-interface TopLevels {
+interface TopLevel {
   @scenario
   @scenarioDoc("""
     Resource GET operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top
     Expected query parameter: api-version=2023-12-01-preview
     
     Expected response body:
     ```json
     {
-      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top",
+      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top",
       "name": "top",
       "type": "topLevel",
       "location": "eastus",
@@ -75,7 +75,7 @@ interface TopLevels {
   @scenario
   @scenarioDoc("""
     Resource PUT operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top
     Expected query parameter: api-version=2023-12-01-preview
     Expected request body:
     ```json
@@ -89,7 +89,7 @@ interface TopLevels {
     Expected response body:
     ```json
     {
-      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top",
+      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top",
       "name": "top",
       "type": "topLevel",
       "location": "eastus",
@@ -113,7 +113,7 @@ interface TopLevels {
   @scenario
   @scenarioDoc("""
     Resource PATCH operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top
     Expected query parameter: api-version=2023-12-01-preview
     Expected request body:
     ```json
@@ -126,7 +126,7 @@ interface TopLevels {
     Expected response body:
     ```json
     {
-      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top",
+      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top",
       "name": "top",
       "type": "topLevel",
       "location": "eastus",
@@ -150,7 +150,7 @@ interface TopLevels {
   @scenario
   @scenarioDoc("""
     Resource DELETE operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top
     Expected query parameter: api-version=2023-12-01-preview
     ```
     Expected response status code: 204
@@ -160,14 +160,14 @@ interface TopLevels {
   @scenario
   @scenarioDoc("""
     Resource LIST by resource group operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources
     Expected query parameter: api-version=2023-12-01-preview
     
     Expected response body:
     ```json
     {
       "value": [{
-        "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top",
+        "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top",
         "name": "top",
         "type": "topLevel",
         "location": "eastus",
@@ -192,14 +192,14 @@ interface TopLevels {
   @scenario
   @scenarioDoc("""
     Resource LIST by subscription operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources
     Expected query parameter: api-version=2023-12-01-preview
     
     Expected response body:
     ```json
     {
       "value": [{
-        "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top",
+        "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top",
         "name": "top",
         "type": "topLevel",
         "location": "eastus",
@@ -224,7 +224,7 @@ interface TopLevels {
   @scenario
   @scenarioDoc("""
       Resource sync action.
-      Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top/actionSync
+      Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resource/topLevelTrackedResources/top/actionSync
       Expected query parameter: api-version=2023-12-01-preview
       Expected request body:
       ```json
